Changes from review - tidying up
Moved structs into their own file
Controlled parallel call with cmake flags
Add parallel libs to link libs
Remove unused headers
